requirements.txt
flask
flask-socketio
flask-wtf
simple-websocket
werkzeug
psycopg2-binary
gunicorn
eventlet
